This will all fall apart anyway

+++ common/src/main/java/li/cil/oc2/common/bus/device/rpc/EnergyStorageDevice.java
@@ -22,12 +22,12 @@ public final class EnergyStorageDevice extends IdentityProxy<EnergyStorage> impl
 
     @Callback
     public int getEnergyStored() {
-        return identity.getEnergyStored();
+        return (int) identity.getEnergyStored();
     }
 
     @Callback
     public int getMaxEnergyStored() {
-        return identity.getMaxEnergyStored();
+        return (int) identity.getMaxEnergyStored();
     }
 
     @Callback
